requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
djangorestframework==3.12.4
greenlet==1.1.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.24
sqlparse==0.4.2
